refactor(casting): rename and redesign media proxy service
Replace IMediaStreamingService/MediaStreamingService with a cleaner
ICastMediaProxy/CastMediaProxy design that better fits the Casting
library's domain language and lifecycle model.

Key changes:
- Rename IMediaStreamingService → ICastMediaProxy and move to
  Abstractions/; the new interface takes IStorageFile directly (not
  object) and returns an ICastProxyHandle instead of managing start/stop
  on a singleton.
- Add ICastProxyHandle : IDisposable (Abstractions/) - a scoped handle
  that owns the StreamSocketListener; disposing it stops the server and
  releases the port automatically.
- Move implementation to Proxy/CastMediaProxy.cs; CastProxyHandle is a
  private inner class, making each StartAsync call fully independent.
  The proxy is now stateless.
- Remove the TryGetNetworkUri dual-responsibility from the proxy; move it
  to CastService as a private static helper. The proxy's only job is now
  proxying local files.
- Add ProxiedCastSession (private inner class in CastService) that wraps
  ICastSession + ICastProxyHandle?, delegating all members and disposing
  both on Dispose(). This ties the proxy lifetime to the session lifetime
  so explicit StopStream() calls are no longer needed.
- Update ServiceHelpers.cs DI registration:
  IMediaStreamingService/MediaStreamingService → ICastMediaProxy/CastMediaProxy
- Delete Services/ folder (was only ever used for MediaStreamingService)
